+++ main.tf
@@ -13,9 +13,9 @@ locals {
   was_missing = var.local_existing_package != null ? !fileexists(var.local_existing_package) : local.archive_was_missing
 
   # s3_* - to get package from S3
-  s3_bucket         = var.s3_existing_package != null ? lookup(var.s3_existing_package, "bucket", null) : (var.store_on_s3 ? var.s3_bucket : null)
-  s3_key            = var.s3_existing_package != null ? lookup(var.s3_existing_package, "key", null) : (var.store_on_s3 ? var.s3_prefix != null ? format("%s%s", var.s3_prefix, replace(local.archive_filename_string, "/^.*//", "")) : replace(local.archive_filename_string, "/^\\.//", "") : null)
-  s3_object_version = var.s3_existing_package != null ? lookup(var.s3_existing_package, "version_id", null) : (var.store_on_s3 ? try(aws_s3_bucket_object.lambda_package[0].version_id, null) : null)
+  s3_bucket         = var.s3_existing_package != null ? try(var.s3_existing_package.bucket, null) : (var.store_on_s3 ? var.s3_bucket : null)
+  s3_key            = var.s3_existing_package != null ? try(var.s3_existing_package.key, null) : (var.store_on_s3 ? var.s3_prefix != null ? format("%s%s", var.s3_prefix, replace(local.archive_filename_string, "/^.*//", "")) : replace(local.archive_filename_string, "/^\\.//", "") : null)
+  s3_object_version = var.s3_existing_package != null ? try(var.s3_existing_package.version_id, null) : (var.store_on_s3 ? try(aws_s3_object.lambda_package[0].version_id, null) : null)
 
 }
 
@@ -37,6 +37,10 @@ resource "aws_lambda_function" "this" {
   package_type                   = var.package_type
   architectures                  = var.architectures
   code_signing_config_arn        = var.code_signing_config_arn
+
+  ephemeral_storage {
+    size = var.ephemeral_storage_size
+  }
 
   filename         = local.filename
   source_code_hash = var.ignore_source_code_hash ? null : (local.filename == null ? false : fileexists(local.filename)) && !local.was_missing ? filebase64sha256(local.filename) : null
@@ -97,7 +101,7 @@ resource "aws_lambda_function" "this" {
   # When a lambda function is invoked, AWS creates the log group automatically if it doesn't exist yet.
   # Without the dependency, this can result in a race condition if the lambda function is invoked before
   # Terraform can create the log group.
-  depends_on = [null_resource.archive, aws_s3_bucket_object.lambda_package, aws_cloudwatch_log_group.lambda]
+  depends_on = [null_resource.archive, aws_s3_object.lambda_package, aws_cloudwatch_log_group.lambda]
 }
 
 resource "aws_lambda_layer_version" "this" {
@@ -118,10 +122,10 @@ resource "aws_lambda_layer_version" "this" {
   s3_key            = local.s3_key
   s3_object_version = local.s3_object_version
 
-  depends_on = [null_resource.archive, aws_s3_bucket_object.lambda_package]
+  depends_on = [null_resource.archive, aws_s3_object.lambda_package]
 }
 
-resource "aws_s3_bucket_object" "lambda_package" {
+resource "aws_s3_object" "lambda_package" {
   count = local.create && var.store_on_s3 && var.create_package ? 1 : 0
 
   bucket        = var.s3_bucket
@@ -201,12 +205,12 @@ resource "aws_lambda_permission" "current_version_triggers" {
   function_name = aws_lambda_function.this[0].function_name
   qualifier     = aws_lambda_function.this[0].version
 
-  statement_id       = lookup(each.value, "statement_id", each.key)
-  action             = lookup(each.value, "action", "lambda:InvokeFunction")
-  principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", null)
-  source_account     = lookup(each.value, "source_account", null)
-  event_source_token = lookup(each.value, "event_source_token", null)
+  statement_id       = try(each.value.statement_id, each.key)
+  action             = try(each.value.action, "lambda:InvokeFunction")
+  principal          = try(each.value.principal, format("%s.amazonaws.com", try(each.value.service, "")))
+  source_arn         = try(each.value.source_arn, null)
+  source_account     = try(each.value.source_account, null)
+  event_source_token = try(each.value.event_source_token, null)
 }
 
 # Error: Error adding new Lambda Permission for lambda: InvalidParameterValueException: We currently do not support adding policies for $LATEST.
@@ -215,12 +219,12 @@ resource "aws_lambda_permission" "unqualified_alias_triggers" {
 
   function_name = aws_lambda_function.this[0].function_name
 
-  statement_id       = lookup(each.value, "statement_id", each.key)
-  action             = lookup(each.value, "action", "lambda:InvokeFunction")
-  principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", null)
-  source_account     = lookup(each.value, "source_account", null)
-  event_source_token = lookup(each.value, "event_source_token", null)
+  statement_id       = try(each.value.statement_id, each.key)
+  action             = try(each.value.action, "lambda:InvokeFunction")
+  principal          = try(each.value.principal, format("%s.amazonaws.com", try(each.value.service, "")))
+  source_arn         = try(each.value.source_arn, null)
+  source_account     = try(each.value.source_account, null)
+  event_source_token = try(each.value.event_source_token, null)
 }
 
 resource "aws_lambda_event_source_mapping" "this" {
@@ -228,23 +232,23 @@ resource "aws_lambda_event_source_mapping" "this" {
 
   function_name = aws_lambda_function.this[0].arn
 
-  event_source_arn = each.value.event_source_arn
+  event_source_arn = try(each.value.event_source_arn, null)
 
-  batch_size                         = lookup(each.value, "batch_size", null)
-  maximum_batching_window_in_seconds = lookup(each.value, "maximum_batching_window_in_seconds", null)
-  enabled                            = lookup(each.value, "enabled", null)
-  starting_position                  = lookup(each.value, "starting_position", null)
-  starting_position_timestamp        = lookup(each.value, "starting_position_timestamp", null)
-  parallelization_factor             = lookup(each.value, "parallelization_factor", null)
-  maximum_retry_attempts             = lookup(each.value, "maximum_retry_attempts", null)
-  maximum_record_age_in_seconds      = lookup(each.value, "maximum_record_age_in_seconds", null)
-  bisect_batch_on_function_error     = lookup(each.value, "bisect_batch_on_function_error", null)
-  topics                             = lookup(each.value, "topics", null)
-  queues                             = lookup(each.value, "queues", null)
-  function_response_types            = lookup(each.value, "function_response_types", null)
+  batch_size                         = try(each.value.batch_size, null)
+  maximum_batching_window_in_seconds = try(each.value.maximum_batching_window_in_seconds, null)
+  enabled                            = try(each.value.enabled, null)
+  starting_position                  = try(each.value.starting_position, null)
+  starting_position_timestamp        = try(each.value.starting_position_timestamp, null)
+  parallelization_factor             = try(each.value.parallelization_factor, null)
+  maximum_retry_attempts             = try(each.value.maximum_retry_attempts, null)
+  maximum_record_age_in_seconds      = try(each.value.maximum_record_age_in_seconds, null)
+  bisect_batch_on_function_error     = try(each.value.bisect_batch_on_function_error, null)
+  topics                             = try(each.value.topics, null)
+  queues                             = try(each.value.queues, null)
+  function_response_types            = try(each.value.function_response_types, null)
 
   dynamic "destination_config" {
-    for_each = lookup(each.value, "destination_arn_on_failure", null) != null ? [true] : []
+    for_each = try(each.value.destination_arn_on_failure, null) != null ? [true] : []
     content {
       on_failure {
         destination_arn = each.value["destination_arn_on_failure"]
@@ -252,8 +256,15 @@ resource "aws_lambda_event_source_mapping" "this" {
     }
   }
 
+  dynamic "self_managed_event_source" {
+    for_each = try(each.value.self_managed_event_source, [])
+    content {
+      endpoints = self_managed_event_source.value.endpoints
+    }
+  }
+
   dynamic "source_access_configuration" {
-    for_each = lookup(each.value, "source_access_configuration", [])
+    for_each = try(each.value.source_access_configuration, [])
     content {
       type = source_access_configuration.value["type"]
       uri  = source_access_configuration.value["uri"]
@@ -261,11 +272,11 @@ resource "aws_lambda_event_source_mapping" "this" {
   }
 
   dynamic "filter_criteria" {
-    for_each = lookup(each.value, "filter_criteria", null) != null ? [true] : []
+    for_each = try(each.value.filter_criteria, null) != null ? [true] : []
 
     content {
       filter {
-        pattern = lookup(each.value["filter_criteria"], "pattern", null)
+        pattern = try(each.value["filter_criteria"].pattern, null)
       }
     }
   }