+++ modules/deploy/versions.tf
@@ -2,8 +2,17 @@ terraform {
   required_version = ">= 0.13.1"
 
   required_providers {
-    aws   = ">= 3.35"
-    local = ">= 1"
-    null  = ">= 2"
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 3.35"
+    }
+    local = {
+      source  = "hashicorp/local"
+      version = ">= 1.0"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = ">= 2.0"
+    }
   }
 }